btn analyze optimized partially
